feat: remove FirebaseDatabaseProvider and FirebaseDatabaseNode from preview
