+++ src/pages/preview/index.tsx
@@ -1,12 +1,45 @@
-import { Preview } from '@/components/Preview'
-import { FirebaseDatabaseProvider } from '@react-firebase/database';
-import firebase from '@/lib/firebase';
+import { useEffect, useState } from 'react';
+import { DataSnapshot } from '@firebase/database-types';
+
+import { db } from '@/lib/firebase';
+import { TextWidget } from '@/components/TextWidget';
+import { TimeWidget } from '@/components/TimeWidget';
+import { IFrameWidget } from '@/components/IFrameWidget';
+
+const Widgets = {
+  'text': TextWidget,
+  'time': TimeWidget,
+  'iframe': IFrameWidget,
+};
+
+type Widget = {
+  name: string;
+  props: any;
+}
+
+type WidgetList = { [key: string]: Widget }
 
 const PreviewPage = () => {
+  const [widgets, setWidgets] = useState<WidgetList>({});
+
+  useEffect(() => {
+    db.ref('/widgets').on('value', (snap: DataSnapshot) => {
+      if (snap?.val()) {
+        setWidgets(snap.val());
+      }
+    });
+  }, [db]);
+
   return (
-    <FirebaseDatabaseProvider firebase={firebase}>
-      <Preview />
-    </FirebaseDatabaseProvider>
+    <div>
+      {
+        Object.keys(widgets).map((id) => {
+          const widget: any = widgets[id];
+          const Widget = Widgets[widget.name];
+          return <Widget key={id} {...widget.props} />
+        })
+      }
+    </div>
   );
 };
